moving print_ast to test executable
